Move script tests to model
