+++ tests/test_models.py
@@ -6,6 +6,7 @@ LICENSE file in the root directory of this source tree.
 """
 
 import pytest
+import torch
 from fastmri.data import transforms
 from fastmri.data.subsample import RandomMaskFunc
 from fastmri.models import Unet, VarNet
@@ -53,3 +54,27 @@ def test_varnet(shape, out_chans, chans, center_fractions, accelerations):
     y = varnet(output, mask.byte())
 
     assert y.shape[1:] == x.shape[2:4]
+
+
+def test_unet_scripting():
+    model = Unet(
+        in_chans=1,
+        out_chans=1,
+        chans=8,
+        num_pool_layers=2,
+        drop_prob=0.0,
+    )
+    scr = torch.jit.script(model)
+    assert scr is not None
+
+
+def test_varnet_scripting():
+    model = VarNet(
+        num_cascades=4,
+        pools=2,
+        chans=8,
+        sens_pools=2,
+        sens_chans=4,
+    )
+    scr = torch.jit.script(model)
+    assert scr is not None
